3 packages, use restore and conditional strategy for cli
