+++ Functions.Templates/Templates/HttpTrigger-CSharp-Isolated/.template.config/template.json
@@ -15,6 +15,21 @@
     },
     "sourceName": "HttpTriggerCSharp",
     "symbols": {
+        "Framework": {
+            "type": "bind",
+            "binding": "msbuild:TargetFramework",
+            "defaultValue": "net6.0"
+        },
+        "NetFramework": {
+            "type": "computed",
+            "datatype": "bool",
+            "value": "(Framework == \"net48\")"
+        },
+        "NetCore": {
+            "type": "computed",
+            "datatype": "bool",
+            "value": "(Framework == \"net6.0\" || Framework == \"net7.0\" || Framework == \"net8.0\")"
+        },
         "namespace": {
             "description": "namespace for the generated code",
             "replaces": "Company.Function",
@@ -46,33 +61,6 @@
             "binding": "HostIdentifier"
         }
     },
-    "Framework": {
-        "type": "bind",
-        "binding": "msbuild:TargetFramework",
-        "defaultValue": "net6.0"
-    },
-    "NetFramework": {
-        "type": "computed",
-        "datatype": "bool",
-        "value": "(Framework == \"net48\")"
-    },
-    "NetCore": {
-        "type": "computed",
-        "datatype": "bool",
-        "value": "(Framework == \"net6.0\" || Framework == \"net7.0\" || Framework == \"net8.0\")"
-    },
-    "sources": [
-        {
-            "condition": "(NetCore == \"True\")",
-            "source": "./",
-            "include": "./HttpTriggerCSharp"
-        },
-        {
-            "condition": "(NetFramework == \"True\")",
-            "source": "./",
-            "include": "./HttpTriggerCSharpNetFx"
-        }
-    ],
     "primaryOutputs": [
         {
             "path": "HttpTriggerCSharp.cs"
@@ -81,7 +69,7 @@
     "defaultName": "HttpTriggerCSharp",
     "postActions": [
         {
-            "condition": "(NetCore == \"True\")",
+            "condition": "(NetCore)",
             "Description": "Adding Reference to Microsoft.Azure.Functions.Worker.Extensions.Http.AspNetCore Nuget package",
             "ActionId": "B17581D1-C5C9-4489-8F0A-004BE667B814",
             "ContinueOnError": "true",
@@ -102,6 +90,12 @@
                 "reference": "Microsoft.Azure.Functions.Worker.Extensions.Http", "version": "3.1.0",
                 "projectFileExtensions": ".csproj"
             }
+        },
+        {
+            "description": "Restore NuGet packages required by this project.",
+            "manualInstructions": [],
+            "actionId": "210D431B-A78B-4D2F-B762-4ED3E3EA9025",
+            "continueOnError": true
         },
         {
             "condition": "(HostIdentifier != \"dotnetcli\" && HostIdentifier != \"dotnetcli-preview\")",
