--- Functions.Templates/Templates/HttpTrigger-CSharp-Isolated/HttpTriggerCSharpNetFx.cs
@@ -1,30 +1,0 @@
-using System.Net;
-using Microsoft.Azure.Functions.Worker;
-using Microsoft.Azure.Functions.Worker.Http;
-using Microsoft.Extensions.Logging;
-
-namespace Company.Function
-{
-    public class HttpTriggerCSharp
-    {
-        private readonly ILogger _logger;
-
-        public HttpTriggerCSharp(ILoggerFactory loggerFactory)
-        {
-            _logger = loggerFactory.CreateLogger<HttpTriggerCSharp>();
-        }
-
-        [Function("HttpTriggerCSharp")]
-        public HttpResponseData Run([HttpTrigger(AuthorizationLevel.AuthLevelValue, "get", "post")] HttpRequestData req)
-        {
-            _logger.LogInformation("C# HTTP trigger function processed a request.");
-
-            var response = req.CreateResponse(HttpStatusCode.OK);
-            response.Headers.Add("Content-Type", "text/plain; charset=utf-8");
-
-            response.WriteString("Welcome to Azure Functions!");
-
-            return response;
-        }
-    }
-}
